requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-PyMongo==2.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==6.0.0
pluggy==0.8.1
py==1.7.0
pymongo==3.7.2
PyYAML==3.13
six==1.12.0
ujson==1.35
urllib3==1.24.1
Werkzeug==0.14.1
gunicorn==19.9.0
redis==3.1.0
